Feature/login

+++ backend/controllers/taskController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const Task = require("../models/taskModel");
+const User = require("../models/userModel");
 
 const getTasks = asyncHandler(async (req, res) => {
-  const tasks = await Task.find();
+  const tasks = await Task.find({ user: req.user.id });
   res.status(200).json(tasks);
 });
 
@@ -11,7 +12,7 @@ const setTask = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Please enter a task");
   }
-  const task = await Task.create({ text: req.body.text });
+  const task = await Task.create({ text: req.body.text, user: req.user.id });
   res.status(200).json(task);
   // res.status(200).json({
   //   message: "Created Task.",
@@ -20,6 +21,7 @@ const setTask = asyncHandler(async (req, res) => {
 
 const updateTask = asyncHandler(async (req, res) => {
   const task = await Task.findById(req.params.id);
+  console.log("task: " + task);
   if (!req.body.text) {
     res.status(400);
     throw new Error("Please enter a task");
@@ -28,13 +30,16 @@ const updateTask = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Task not found");
   }
+
+  if (task.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User is not authorized to update");
+  }
+
   const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
   res.status(200).json(updatedTask);
-  // res.status(200).json({
-  //   message: `Task ${req.params.id} updated.`,
-  // });
 });
 
 const deleteTask = asyncHandler(async (req, res) => {
@@ -44,12 +49,13 @@ const deleteTask = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Task not found");
   }
+  if (task.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User is Not authorized to delete task");
+  }
   await Task.findByIdAndDelete(req.params.id);
 
   res.status(200).json({ id: req.params.id });
-  // res.status(200).json({
-  //   message: `Task ${req.params.id} deleted.`,
-  // });
 });
 
 module.exports = { getTasks, setTask, updateTask, deleteTask };
